fix comment

+++ src/main/java/com/midorimart/managementsystem/model/product/dto/ProductDTOFilter.java
@@ -21,5 +21,5 @@ public class ProductDTOFilter {
     private int offset;
     private int merchantId;
     private List<String> origin;
-    private List<Integer> star;
+    private List<Double> star;
 }
